Run with py3.10 and migrate tests
