+++ salt/transport/base.py
@@ -3,6 +3,7 @@ import os
 TRANSPORTS = (
     "zeromq",
     "tcp",
+    "ws",
 )
 
 
@@ -24,6 +25,10 @@ def request_server(opts, **kwargs):
         import salt.transport.tcp
 
         return salt.transport.tcp.TCPReqServer(opts)
+    elif ttype == "ws":
+        import salt.transport.ws
+
+        return salt.transport.ws.RequestServer(opts)
     elif ttype == "local":
         import salt.transport.local
 
@@ -46,6 +51,10 @@ def request_client(opts, io_loop):
         import salt.transport.tcp
 
         return salt.transport.tcp.TCPReqClient(opts, io_loop=io_loop)
+    elif ttype == "ws":
+        import salt.transport.ws
+
+        return salt.transport.ws.RequestClient(opts, io_loop=io_loop)
     else:
         raise Exception("Channels are only defined for tcp, zeromq")
 
@@ -80,6 +89,10 @@ def publish_server(opts, **kwargs):
         import salt.transport.tcp
 
         return salt.transport.tcp.TCPPublishServer(opts, **kwargs)
+    elif ttype == "ws":
+        import salt.transport.ws
+
+        return salt.transport.ws.PublishServer(opts, **kwargs)
     elif ttype == "local":  # TODO:
         import salt.transport.local
 
@@ -126,6 +139,12 @@ def publish_client(opts, io_loop, host=None, port=None, path=None):
         import salt.transport.tcp
 
         return salt.transport.tcp.TCPPubClient(
+            opts, io_loop, host=host, port=port, path=path
+        )
+    elif ttype == "ws":
+        import salt.transport.ws
+
+        return salt.transport.ws.PublishClient(
             opts, io_loop, host=host, port=port, path=path
         )
     raise Exception("Transport type not found: {}".format(ttype))
